backend/requirements.txt
fastapi==0.115.3
uvicorn==0.32.0
httpx==0.27.2
redis==5.1.1
setuptools==75.2.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib==1.7.4
mysql-connector-python==9.1.0
sqlmodel==0.0.22
PyJWT==2.9.0
pydantic==2.9.2